refactor(hooks): Move useMarketDataViewsConfig to shared hooks
