+++ js/src/pages/markets/markets-dashboard/market-data-views/index.js
@@ -10,11 +10,10 @@ import { edit, trash } from '@wordpress/icons';
  * Internal dependencies
  */
 import { PRIMARY_MARKET_ID } from '../../constants';
-import useMarkets from '~/hooks/useMarkets';
 import useTargetAudienceFinalCountryCodes from '~/hooks/useTargetAudienceFinalCountryCodes';
+import useMarketDataViewsConfig from '~/hooks/useMarketDataViewsConfig';
 import EditMarketModal from '../edit-market-modal';
 import DeleteMarketModal from '../delete-market-modal';
-import useMarketDataViewsConfig from './use-market-data-views-config';
 import './index.scss';
 
 const isPrimaryMarket = ( market ) => market.id === PRIMARY_MARKET_ID;
@@ -34,8 +33,7 @@ const DEFAULT_VIEW = {
  */
 const MarketDataViews = () => {
 	const { DataViews } = window.wp.dataviews;
-	const { fields, data } = useMarketDataViewsConfig();
-	const { hasFinishedResolution } = useMarkets();
+	const { fields, data, hasFinishedResolution } = useMarketDataViewsConfig();
 	const [ view, setView ] = useState( DEFAULT_VIEW );
 	const [ editingMarket, setEditingMarket ] = useState( null );
 	const [ deletingMarket, setDeletingMarket ] = useState( null );
